pyproject.toml
[tool.poetry]
name = "rzepabot"
version = "0.1.0"
description = "Polski bot dyskordowy Animal Crossing"
authors = ["Slavfox <slavfoxman@gmail.com>"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.2"
peewee = "^3.13.2"
matplotlib = "^3.2.1"
pendulum = "^2.1.0"

[tool.poetry.dev-dependencies]
lxml = "^4.5.0"

[tool.black]
line-length = 79

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79
known_third_party = ["snecs"]
force_to_top = ["typing"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
